app/requirements.txt
aiosmtplib==2.0.1
alembic==1.10.2
anyio==3.6.2
bcrypt==4.0.1
blinker==1.6
bson==0.5.10
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
fastapi-mail==1.2.7
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
Pillow==9.2.0
numpy==1.23.4
torch==1.12.1
onnx==1.12.0
torchvision==0.13.1
onnxruntime==1.12.1